add&feat : BaseActivity 구현 성공!!! #8
