+++ app/src/main/AndroidManifest.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools" >
+    xmlns:tools="http://schemas.android.com/tools">
     <!-- 갤러리 권한 -->
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
+    <uses-permission android:name="android.permission.INTERNET" />
 
-    <uses-permission android:name="android.permission.INTERNET"/>
     <uses-feature
         android:name="android.hardware.camera"
         android:required="true" />
@@ -16,7 +16,6 @@
         </intent>
     </queries> <!-- 갤러리 권한 -->
     <uses-permission android:name="android.permission.CAMERA" />
-
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
     <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION" />
@@ -32,11 +31,22 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.EatSSUAndroid"
         android:usesCleartextTraffic="true"
-        tools:targetApi="31" >
-        
+        tools:targetApi="31">
+        <activity
+            android:name=".ui.BaseActivity"
+            android:exported="false">
+            <meta-data
+                android:name="android.app.lib_name"
+                android:value="" />
+        </activity>
         <activity
             android:name=".ui.login.SignInActivity"
-            android:exported="false" >
+            android:exported="true">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
@@ -44,7 +54,7 @@
         <activity
             android:name=".ui.mypage.ChangePwActivity"
             android:exported="false"
-            android:windowSoftInputMode="adjustResize" >
+            android:windowSoftInputMode="adjustResize">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
@@ -52,7 +62,7 @@
         <activity
             android:name=".ui.mypage.ChangeNicknameActivity"
             android:exported="false"
-            android:windowSoftInputMode="adjustResize" >
+            android:windowSoftInputMode="adjustResize">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
@@ -60,15 +70,14 @@
         <activity
             android:name=".ui.login.LoginActivity"
             android:exported="false"
-            android:windowSoftInputMode="adjustResize" >
+            android:windowSoftInputMode="adjustResize">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
-
         <activity
             android:name=".ui.mypage.MyReviewListActivity"
-            android:exported="true" >
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -81,7 +90,7 @@
         </activity>
         <activity
             android:name=".ui.review.ReviewListActivity"
-            android:exported="true" >
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -94,7 +103,7 @@
         </activity>
         <activity
             android:name=".ui.mypage.MyPageActivity"
-            android:exported="true" >
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -108,7 +117,7 @@
         <activity
             android:name=".ui.review.WriteReview2Activity"
             android:exported="false"
-            android:windowSoftInputMode="adjustResize" >
+            android:windowSoftInputMode="adjustResize">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
@@ -116,7 +125,7 @@
         <activity
             android:name=".ui.review.WriteReview1Activity"
             android:exported="false"
-            android:windowSoftInputMode="adjustResize" >
+            android:windowSoftInputMode="adjustResize">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
@@ -125,7 +134,6 @@
             android:name=".InfoActivity_Gisik"
             android:exported="true"
             android:theme="@style/Theme.AppCompat.Dialog">
-
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
@@ -134,7 +142,6 @@
             android:name=".InfoActivity_Kitchen"
             android:exported="true"
             android:theme="@style/Theme.AppCompat.Dialog">
-
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
@@ -143,7 +150,6 @@
             android:name=".InfoActivity_Food"
             android:exported="true"
             android:theme="@style/Theme.AppCompat.Dialog">
-
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
@@ -152,7 +158,6 @@
             android:name=".InfoActivity_Snack"
             android:exported="true"
             android:theme="@style/Theme.AppCompat.Dialog">
-
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
@@ -161,7 +166,6 @@
             android:name=".InfoActivity_Dodam"
             android:exported="true"
             android:theme="@style/Theme.AppCompat.Dialog">
-
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
@@ -194,7 +198,7 @@
         </activity>
         <activity
             android:name=".MainActivity"
-            android:exported="true" >
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
